+++ packages/dash-table/src/core/environment/index.ts
@@ -6,14 +6,14 @@ import { Edge } from 'dash-table/derived/edges/type';
 const DASH_DEBUG = 'dash_debug';
 const DASH_LOG = 'dash_log';
 
-const DEFAULT_EDGE: Edge = '1px solid #d3d3d3';
-const ACTIVE_EDGE: Edge = '1px solid var(--accent)';
-
 interface ISearchParams {
     get: (key: string) => string | null;
 }
 
 export default class Environment {
+    private static readonly _supportsCssVariables = Boolean(window.CSS?.supports?.('.some-selector', 'var(--some-var)'));
+    private static readonly _activeEdge: Edge = Environment._supportsCssVariables ? '1px solid var(--accent)' : '1px solid hotpink';
+
     public static get searchParams(): ISearchParams {
         return (
             typeof URL !== 'undefined' &&
@@ -40,10 +40,14 @@ export default class Environment {
     }
 
     public static get defaultEdge(): Edge {
-        return DEFAULT_EDGE;
+        return '1px solid #d3d3d3';
     }
 
     public static get activeEdge(): Edge {
-        return ACTIVE_EDGE;
+        return Environment._activeEdge;
+    }
+
+    public static get supportsCssVariables(): boolean {
+        return Environment._supportsCssVariables;
     }
 }