+++ packages/dash-table/src/dash-table/dash/DataTable.js
@@ -108,7 +108,7 @@ export const defaultProps = {
 
 export const propTypes = {
     /**
-     * The row and column indices and IDs of the currently active cell. 
+     * The row and column indices and IDs of the currently active cell.
      * `row_id` is only returned if the data rows have an `id` key.
      */
     active_cell: PropTypes.exact({
@@ -1077,7 +1077,7 @@ export const propTypes = {
      */
     style_cell_conditional: PropTypes.arrayOf(PropTypes.shape({
         if: PropTypes.exact({
-            column_id: PropTypes.string,
+            column_id: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
             column_type: PropTypes.oneOf(['any', 'numeric', 'text', 'datetime'])
         })
     })),
@@ -1088,12 +1088,14 @@ export const propTypes = {
      */
     style_data_conditional: PropTypes.arrayOf(PropTypes.shape({
         if: PropTypes.exact({
-            column_id: PropTypes.string,
+            column_id: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
             column_type: PropTypes.oneOf(['any', 'numeric', 'text', 'datetime']),
             filter_query: PropTypes.string,
+            state: PropTypes.oneOf(['active', 'selected']),
             row_index: PropTypes.oneOfType([
                 PropTypes.number,
-                PropTypes.oneOf(['odd', 'even'])
+                PropTypes.oneOf(['odd', 'even']),
+                PropTypes.arrayOf(PropTypes.number)
             ]),
             column_editable: PropTypes.bool
         })
@@ -1105,7 +1107,7 @@ export const propTypes = {
      */
     style_filter_conditional: PropTypes.arrayOf(PropTypes.shape({
         if: PropTypes.exact({
-            column_id: PropTypes.string,
+            column_id: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
             column_type: PropTypes.oneOf(['any', 'numeric', 'text', 'datetime']),
             column_editable: PropTypes.bool
         })
@@ -1117,10 +1119,11 @@ export const propTypes = {
      */
     style_header_conditional: PropTypes.arrayOf(PropTypes.shape({
         if: PropTypes.exact({
-            column_id: PropTypes.string,
+            column_id: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
             column_type: PropTypes.oneOf(['any', 'numeric', 'text', 'datetime']),
             header_index: PropTypes.oneOfType([
                 PropTypes.number,
+                PropTypes.arrayOf(PropTypes.number),
                 PropTypes.oneOf(['odd', 'even'])
             ]),
             column_editable: PropTypes.bool