+++ packages/dash-table/src/dash-table/derived/cell/wrapperStyles.ts
@@ -2,11 +2,16 @@ import * as R from 'ramda';
 import { CSSProperties } from 'react';
 
 import { memoizeOneFactory } from 'core/memoizer';
-import { Data, Columns, IViewportOffset, SelectedCells } from 'dash-table/components/Table/props';
+import { Data, Columns, IViewportOffset, SelectedCells, ICellCoordinates } from 'dash-table/components/Table/props';
 import { IConvertedStyle, getDataCellStyle, getDataOpCellStyle } from '../style';
 import { traverseMap2, shallowClone } from 'core/math/matrixZipMap';
 
-const SELECTED_CELL_STYLE = { backgroundColor: 'var(--selected-background)' };
+import isActiveCell from 'dash-table/derived/cell/isActive';
+import Environment from 'core/environment';
+
+const SELECTED_CELL_STYLE = {
+    backgroundColor: Environment.supportsCssVariables ? 'var(--selected-background)' : 'rgba(255, 65, 54, 0.2)'
+};
 
 const partialGetter = (
     columns: Columns,
@@ -16,28 +21,49 @@ const partialGetter = (
 ) => traverseMap2(
     data,
     columns,
-    (datum, column, i) => getDataCellStyle(datum, i + offset.rows, column)(styles)
+    (datum, column, i) => getDataCellStyle(datum, i + offset.rows, column, false, false)(styles)
 );
 
 const getter = (
-    styles: CSSProperties[][],
+    baseline: CSSProperties[][],
+    columns: Columns,
+    styles: IConvertedStyle[],
+    data: Data,
     offset: IViewportOffset,
+    activeCell: ICellCoordinates,
     selectedCells: SelectedCells
 ) => {
-    styles = shallowClone(styles);
+    baseline = shallowClone(baseline);
+
+    const cells = selectedCells.length ?
+        selectedCells :
+        activeCell ? [activeCell] : [];
+
+    const inactiveStyles = styles.filter(style => !style.checksState());
+    const selectedStyles = styles.filter(style => style.checksStateSelected());
+    const activeStyles = styles.filter(style => style.checksStateActive());
 
     R.forEach(({ row: i, column: j }) => {
-        i -= offset.rows;
-        j -= offset.columns;
+        const iNoOffset = i - offset.rows;
+        const jNoOffset = j - offset.columns;
 
-        if (i < 0 || j < 0 || styles.length <= i || styles[i].length <= j) {
+        if (iNoOffset < 0 || jNoOffset < 0 || baseline.length <= iNoOffset || baseline[iNoOffset].length <= jNoOffset) {
             return;
         }
 
-        styles[i][j] = R.merge(styles[i][j], SELECTED_CELL_STYLE);
-    }, selectedCells);
+        const active = isActiveCell(activeCell, i, j);
 
-    return styles;
+        const style = {
+            ...getDataCellStyle(data[i], i + offset.rows, columns[j], active, true)(inactiveStyles),
+            ...SELECTED_CELL_STYLE,
+            ...getDataCellStyle(data[i], i + offset.rows, columns[j], active, true)(selectedStyles),
+            ...getDataCellStyle(data[i], i + offset.rows, columns[j], active, true)(activeStyles)
+        }
+
+        baseline[iNoOffset][jNoOffset] = style;
+    }, cells);
+
+    return baseline;
 };
 
 const opGetter = (